+++ src/main/java/dev/demonicrous/furusato/api/IFurusatoAPI.java
@@ -1,6 +1,7 @@
 package dev.demonicrous.furusato.api;
 
 import dev.demonicrous.furusato.api.bootstrap.BootstrapReport;
+import dev.demonicrous.furusato.api.module.IModuleManager;
 import dev.demonicrous.furusato.api.service.IServiceRegistry;
 
 public interface IFurusatoAPI {
@@ -11,6 +12,8 @@ public interface IFurusatoAPI {
     CoreState state();
 
     BootstrapReport bootstrapReport();
+
+    IModuleManager modules();
 
     IServiceRegistry services();
 }