+++ src/main/java/dev/demonicrous/furusato/api/module/IFurusatoModule.java
@@ -1,0 +1,11 @@
+package dev.demonicrous.furusato.api.module;
+
+public interface IFurusatoModule {
+    ModuleMetadata metadata();
+
+    void onLoad(ModuleContext context) throws Exception;
+
+    void onEnable() throws Exception;
+
+    void onDisable() throws Exception;
+}