+++ src/main/java/dev/demonicrous/furusato/api/module/IModuleManager.java
@@ -1,0 +1,12 @@
+package dev.demonicrous.furusato.api.module;
+
+import java.util.Collection;
+import java.util.Optional;
+
+public interface IModuleManager {
+    void register(IFurusatoModule module);
+
+    Optional<ModuleContainer> find(String moduleId);
+
+    Collection<ModuleContainer> containers();
+}